Improve jcenter support with the plugin keyword
Adds lazy evaluation all the way down to enable early declaration of the
plugin. It doesn't require to be applied after the project's build
configuration anymore.

A dedicated class was added for the task to make things easier, but no
extension class for customization.

The task was renamed to just 'syntastic', but it didn't break the
integration tests, they don't seem to run.